refactor(routing): move router test plugins to testplugins.router sub-package
Partially addresses the busy testplugins package (k-wall comment r3552729187)
by extracting the 7 RouterFactory implementations into
io.kroxylicious.it.testplugins.router. Filter implementations remain in
io.kroxylicious.it.testplugins. Adds package-info.java for the new package.
