+++ kroxylicious-integration-tests/src/test/java/io/kroxylicious/it/RouterChangeHotReloadIT.java
@@ -22,7 +22,7 @@ import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.kroxylicious.it.testplugins.InvocationCountingRouterFactory;
+import io.kroxylicious.it.testplugins.router.InvocationCountingRouterFactory;
 import io.kroxylicious.proxy.config.ClusterDefinition;
 import io.kroxylicious.proxy.config.RouteDefinition;
 import io.kroxylicious.proxy.config.RouteTarget;
