+++ kroxylicious-integration-tests/src/test/java/io/kroxylicious/it/RoutingMultiBackendIT.java
@@ -20,7 +20,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 
 import io.github.nettyplus.leakdetector.junit.NettyLeakDetectorExtension;
 
-import io.kroxylicious.it.testplugins.AlternatingRouterFactory;
+import io.kroxylicious.it.testplugins.router.AlternatingRouterFactory;
 import io.kroxylicious.proxy.config.ClusterDefinition;
 import io.kroxylicious.proxy.config.ConfigurationBuilder;
 import io.kroxylicious.proxy.config.RouteDefinition;
