+++ kroxylicious-integration-tests/src/test/java/io/kroxylicious/it/RoutingTlsIT.java
@@ -24,7 +24,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 
 import io.github.nettyplus.leakdetector.junit.NettyLeakDetectorExtension;
 
-import io.kroxylicious.it.testplugins.PassThroughRouterFactory;
+import io.kroxylicious.it.testplugins.router.PassThroughRouterFactory;
 import io.kroxylicious.proxy.config.ClusterDefinition;
 import io.kroxylicious.proxy.config.ConfigurationBuilder;
 import io.kroxylicious.proxy.config.RouteDefinition;
