+++ kroxylicious-integration-tests/src/test/java/io/kroxylicious/it/testplugins/router/AlternatingRouterFactory.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.kroxylicious.it.testplugins;
+package io.kroxylicious.it.testplugins.router;
 
 import java.util.Arrays;
 import java.util.Map;
