+++ kroxylicious-integration-tests/src/test/resources/META-INF/services/io.kroxylicious.proxy.router.RouterFactory
@@ -3,10 +3,10 @@
 #
 # Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
 #
-io.kroxylicious.it.testplugins.AlternatingRouterFactory
-io.kroxylicious.it.testplugins.ContextCapturingRouterFactory
-io.kroxylicious.it.testplugins.DynamicProduceRouterFactory
-io.kroxylicious.it.testplugins.InvocationCountingRouterFactory
-io.kroxylicious.it.testplugins.PassThroughRouterFactory
-io.kroxylicious.it.testplugins.RouterContractCapturingFactory
-io.kroxylicious.it.testplugins.SplitStaticRouterFactory
+io.kroxylicious.it.testplugins.router.AlternatingRouterFactory
+io.kroxylicious.it.testplugins.router.ContextCapturingRouterFactory
+io.kroxylicious.it.testplugins.router.DynamicProduceRouterFactory
+io.kroxylicious.it.testplugins.router.InvocationCountingRouterFactory
+io.kroxylicious.it.testplugins.router.PassThroughRouterFactory
+io.kroxylicious.it.testplugins.router.RouterContractCapturingFactory
+io.kroxylicious.it.testplugins.router.SplitStaticRouterFactory
